dom/media/mediasource/test/test_DurationUpdated_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: append data and check that mediasource duration got updated</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/mp4");

  let durationChangeCount = 0;
  v.addEventListener("durationchange", () => durationChangeCount++);

  const arrayBuffer = await fetchWithXHR("bipbop/bipbop2s.mp4");
  sb.appendBuffer(new Uint8Array(arrayBuffer, 0, 1395));

  // Adding the first init segment will fire a durationchange.
  await Promise.all([once(sb, "updateend"), once(v, "loadedmetadata")]);
  ok(true, "got loadedmetadata");
  // Set mediasource duration to 0, so future appendBuffer
  // will update the mediasource duration.
  // Changing the duration will fire a durationchange.
  ms.duration = 0;
  sb.appendBuffer(new Uint8Array(arrayBuffer, 1395));
  // Adding more data will fire durationchange.
  await once(sb, "updateend");
  ok(true, "got updateend");
  // this will not fire durationchange as new duration == old duration
  ms.endOfStream();
  await once(ms, "sourceended");
  is(durationChangeCount, 3, "durationchange not fired as many times as expected");
  is(v.duration, 1.696666, "Video has correct duration");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
